refactor all things using custom hooks
