+++ src/hooks/useNow.ts
@@ -1,0 +1,18 @@
+import dayjs from 'dayjs'
+import React from 'react'
+
+export function useNow(enabled: boolean) {
+  const [now, setNow] = React.useState(dayjs())
+
+  React.useEffect(() => {
+    if (!enabled) {
+      return () => {}
+    }
+    const pid = setInterval(() => setNow(dayjs()), 2 * 60 * 1000)
+    return () => clearInterval(pid)
+  }, [enabled])
+
+  return {
+    now,
+  }
+}
